MyMessages web application using MEAN stack.
These are the files for the web application that i made using MEAN stack web application where the user can add a new post and with an image addition feature along with content. 
The frontend is taken care of by Angular( Html, CSS, Typescript ) and the backend is done my NodeJs, ExpressJs and the no-sql  database is from MongoDb.